move other generate path to hook
